FinSight-Agent/requirements.txt
langgraph
langchain
langchain_community
python-dotenv
langchain-groq
langchain_tavily
yfinance
streamlit